requirements.txt
# Core dependencies
web3==6.0.0
aiohttp==3.8.4
pyyaml==6.0
loguru==0.7.0
python-dotenv==1.0.0
asyncio==3.4.3
nest-asyncio==1.5.6

# Blockchain specific
eth-account==0.8.0
eth-typing==3.3.0
eth-utils==2.1.0
solana==0.29.0
solders==0.18.0
anchorpy==0.17.0

# Utilities
pydantic==1.10.7
tenacity==8.2.2
cachetools==5.3.0
cryptography==40.0.2
python-dateutil==2.8.2
statistics==1.0.3.5

# Development
pytest==7.3.1
pytest-asyncio==0.21.0
black==23.3.0
isort==5.12.0
mypy==1.3.0
flake8==6.0.0

# Dépendances principales
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.3.0
scikit-learn>=1.0.0

# Base de données
sqlalchemy>=1.4.0
asyncpg>=0.25.0
psycopg2-binary>=2.9.0

# Web et API
flask>=2.0.0
flask-cors>=3.0.10
flask-limiter>=2.8.0
websockets>=10.3.0

# Monitoring et sécurité
prometheus-client>=0.14.0
pyOpenSSL>=22.0.0
certifi>=2022.5.18

# Nouvelles dépendances
requests>=2.26.0
scipy==1.10.1

# Solana dependencies
solana-py==0.30.2
solders-macros==0.2.0

# General dependencies
requests>=2.28.1
aiohttp>=3.8.3
python-dotenv>=0.21.0
websockets>=10.4
numpy>=1.24.1
pandas>=1.5.2

# Specific to the project
pytest>=7.2.0
pytest-asyncio>=0.20.3